Separate search from runtime
