+++ crates/wasvy_cli/src/runtime.rs
@@ -1,13 +1,7 @@
-use std::{
-    collections::{HashMap, HashSet},
-    iter,
-    path::{Path, PathBuf},
-    sync::Arc,
-};
+use std::{collections::HashMap, iter, path::Path, sync::Arc};
 
 use anyhow::{Context, Result, bail};
 use error_collection::Errors;
-use serde::Deserialize;
 use wit_parser::Resolve;
 
 use crate::{
@@ -16,9 +10,9 @@ use crate::{
     editor::BoxedEditor,
     id::Id,
     language::BoxedLanguage,
-    languages::{Rust, cargo_metadata},
     named::Named,
     remote::Remote,
+    search::SearchBuilder,
     source::Source,
 };
 
@@ -193,53 +187,11 @@ impl Runtime {
     /// - External mods located somewhere within the path (Rust, Python, Go, etc)
     /// - Pre-compiled binaries located somewhere within the path (Wasm)
     pub fn search(&self, remote: &Remote, path: impl AsRef<Path>) -> Result<Vec<Source>> {
-        let Native {
-            crate_names,
-            workspace_root,
-        } = find_native(&remote.current_exe);
-        let native: Vec<Source> = crate_names
-            .into_iter()
-            .filter_map(|crate_name| Source::new_native(&workspace_root, crate_name, self).ok())
-            .collect();
-
-        let rust = Rust::id();
-        let mut mods: Vec<Source> = search_glob(path.as_ref().join("**/wit/*.wit"))
-            .filter_map(|p| p.parent().and_then(Path::parent).map(Path::to_path_buf))
-            .collect::<HashSet<_>>() // Dedupe
-            .into_iter()
-            .filter_map(|path| Source::new_dir(&path, self).ok())
-            // Avoid duplicates with native sources
-            .filter(|source| {
-                !(source.path().starts_with(&workspace_root)
-                    && source.is_language(&rust)
-                    && native.iter().any(|native| native.name() == source.name()))
-            })
-            .collect();
-
-        let mut wasm: Vec<Source> = search_glob(path.as_ref().join("**/*.wasm"))
-            // Ignore wasm build artifacts located in source directories (such as dest directory for python)
-            .filter(|path| !mods.iter().any(|source| path.starts_with(source.path())))
-            // Ignore wasm files in rust build directory (**/target/wasm32-*/*/*.wasm )
-            .filter(|path| {
-                let mut components = path.components().rev().skip(2);
-                let target = components
-                    .next()
-                    .and_then(|part| part.as_os_str().to_str())
-                    .map(|part| part.starts_with("wasm32-"))
-                    .unwrap_or_default();
-                let dir = components
-                    .next()
-                    .map(|part| part.as_os_str() == "target")
-                    .unwrap_or_default();
-                !target || !dir
-            })
-            .filter_map(|path| Source::new_wasm(path, None, self).ok())
-            .collect();
-
-        let mut sources = native;
-        sources.append(&mut mods);
-        sources.append(&mut wasm);
-        Ok(sources)
+        SearchBuilder::new(self)
+            .native(&remote.current_exe)
+            .dir(path.as_ref())
+            .wasm(path.as_ref())
+            .search()
     }
 
     /// Identifies a directory directory as a compatible [Source] (build files) for a Mod
@@ -256,66 +208,5 @@ impl Runtime {
         logging: Logging,
     ) -> Result<Source> {
         Source::scaffold(name, path, self, language, logging)
-    }
-}
-
-fn search_glob(pattern: PathBuf) -> impl Iterator<Item = PathBuf> {
-    let pattern = pattern.to_str().expect("unicode path");
-    glob::glob(pattern)
-        .expect("valid glob")
-        .filter_map(Result::ok)
-        .filter(|path| {
-            // Omit hidden directories
-            !path
-                .components()
-                .any(|component| component.as_os_str().to_string_lossy().starts_with('.'))
-        })
-}
-
-#[derive(Deserialize, Default)]
-struct Native {
-    crate_names: Vec<String>,
-    workspace_root: PathBuf,
-}
-
-fn find_native(path: impl AsRef<Path>) -> Native {
-    #[derive(Deserialize, Default)]
-    struct Metadata {
-        packages: Vec<Package>,
-        workspace_root: PathBuf,
-    }
-
-    #[derive(Deserialize, Default)]
-    struct Package {
-        name: String,
-        targets: Vec<Target>,
-    }
-
-    #[derive(Deserialize, Default)]
-    struct Target {
-        crate_types: HashSet<String>,
-    }
-
-    let metadata = cargo_metadata(path).unwrap_or_default();
-    let Metadata {
-        packages,
-        workspace_root,
-    } = serde_json::from_str(&metadata).unwrap_or_default();
-
-    let crate_names = packages
-        .into_iter()
-        .filter(|package| {
-            package.targets.iter().any(|target| {
-                // All native mods have these two crate types.
-                // TODO: there's probably more we can check here to avoid false positives
-                target.crate_types.contains("rlib") && target.crate_types.contains("cdylib")
-            })
-        })
-        .map(|package| package.name)
-        .collect();
-
-    Native {
-        crate_names,
-        workspace_root,
     }
 }
